# openapi.yaml
...
openapi: 3.1.0
info:
  title: photo-portfolio-store
  version: 0.1.0
  description: |
    Photographer portfolio + print storefront API.

    Bounded contexts: auth, events, media, search, catalog, commerce,
    compliance, admin. Each operation declares its context via tag.
servers:
  - url: http://localhost:4000
    description: local
...
security: []
paths:
...
  /v1/auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new user
      x-status: stub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created and authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Email/password login
      x-status: stub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/auth/refresh:
    post:
      operationId: authRefresh
      tags: [auth]
      summary: Exchange refresh token for a new access token
      x-status: stub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New tokens issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Revoke the current refresh session
      x-status: stub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Session revoked.
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Conflict:
      description: Conflict with current resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            trace_id:
              type: string
...
    UserRole:
      type: string
      enum: [superadmin, admin, photographer, organizer, assistant, attendee]
    UserStatus:
      type: string
      enum: [active, suspended, deleted]
    User:
      type: object
      required: [id, email, role, status, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        email: { type: string, format: email }
        displayName: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        createdAt: { type: string, format: date-time }
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        displayName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
...
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn:
          type: integer
          description: Access-token TTL in seconds.
        user: { $ref: '#/components/schemas/User' }
...